+++ cloud/functions/src/utils/resources.ts
@@ -3,7 +3,7 @@ export class Resources {
     static get Constants(): any {
         return {
             PRESENTATION: {
-                URL: 'https://deckdeckgo-beta-presentations.deckdeckgo.io',
+                URL: 'https://beta.deckdeckgo.io',
                 FOLDER: 'presentations',
                 IMAGE: 'deckdeckgo.png'
             }