+++ studio/src/components.d.ts
@@ -38,6 +38,7 @@ export namespace Components {
   interface AppContact {}
   interface AppCreateSlide {}
   interface AppCustomImages {}
+  interface AppDashboard {}
   interface AppDeckOrSlide {
     'deckOrSlide': boolean;
   }
@@ -56,6 +57,7 @@ export namespace Components {
     'touch': (element: HTMLElement) => Promise<void>;
     'unSelect': () => Promise<void>;
   }
+  interface AppFaq {}
   interface AppFeed {}
   interface AppFeedCard {
     'compact': boolean;
@@ -190,6 +192,12 @@ declare global {
     new (): HTMLAppCustomImagesElement;
   };
 
+  interface HTMLAppDashboardElement extends Components.AppDashboard, HTMLStencilElement {}
+  var HTMLAppDashboardElement: {
+    prototype: HTMLAppDashboardElement;
+    new (): HTMLAppDashboardElement;
+  };
+
   interface HTMLAppDeckOrSlideElement extends Components.AppDeckOrSlide, HTMLStencilElement {}
   var HTMLAppDeckOrSlideElement: {
     prototype: HTMLAppDeckOrSlideElement;
@@ -218,6 +226,12 @@ declare global {
   var HTMLAppEditorToolbarElement: {
     prototype: HTMLAppEditorToolbarElement;
     new (): HTMLAppEditorToolbarElement;
+  };
+
+  interface HTMLAppFaqElement extends Components.AppFaq, HTMLStencilElement {}
+  var HTMLAppFaqElement: {
+    prototype: HTMLAppFaqElement;
+    new (): HTMLAppFaqElement;
   };
 
   interface HTMLAppFeedElement extends Components.AppFeed, HTMLStencilElement {}
@@ -480,11 +494,13 @@ declare global {
     'app-contact': HTMLAppContactElement;
     'app-create-slide': HTMLAppCreateSlideElement;
     'app-custom-images': HTMLAppCustomImagesElement;
+    'app-dashboard': HTMLAppDashboardElement;
     'app-deck-or-slide': HTMLAppDeckOrSlideElement;
     'app-developer': HTMLAppDeveloperElement;
     'app-editor': HTMLAppEditorElement;
     'app-editor-actions': HTMLAppEditorActionsElement;
     'app-editor-toolbar': HTMLAppEditorToolbarElement;
+    'app-faq': HTMLAppFaqElement;
     'app-feed': HTMLAppFeedElement;
     'app-feed-card': HTMLAppFeedCardElement;
     'app-feed-card-tags': HTMLAppFeedCardTagsElement;
@@ -546,7 +562,7 @@ declare namespace LocalJSX {
     'background'?: string;
     'color'?: string;
     'deckOrSlide'?: boolean;
-    'onColorDidChange'?: (event: CustomEvent<HTMLElement>) => void;
+    'onColorDidChange'?: (event: CustomEvent<boolean>) => void;
     'selectedElement'?: HTMLElement;
   }
   interface AppContact extends JSXBase.HTMLAttributes<HTMLAppContactElement> {}
@@ -554,6 +570,7 @@ declare namespace LocalJSX {
     'onSignIn'?: (event: CustomEvent<void>) => void;
   }
   interface AppCustomImages extends JSXBase.HTMLAttributes<HTMLAppCustomImagesElement> {}
+  interface AppDashboard extends JSXBase.HTMLAttributes<HTMLAppDashboardElement> {}
   interface AppDeckOrSlide extends JSXBase.HTMLAttributes<HTMLAppDeckOrSlideElement> {
     'deckOrSlide'?: boolean;
     'onApplyTo'?: (event: CustomEvent<boolean>) => void;
@@ -582,6 +599,7 @@ declare namespace LocalJSX {
     'onSlideDelete'?: (event: CustomEvent<HTMLElement>) => void;
     'onSlideDidChange'?: (event: CustomEvent<HTMLElement>) => void;
   }
+  interface AppFaq extends JSXBase.HTMLAttributes<HTMLAppFaqElement> {}
   interface AppFeed extends JSXBase.HTMLAttributes<HTMLAppFeedElement> {}
   interface AppFeedCard extends JSXBase.HTMLAttributes<HTMLAppFeedCardElement> {
     'compact'?: boolean;
@@ -684,11 +702,13 @@ declare namespace LocalJSX {
     'app-contact': AppContact;
     'app-create-slide': AppCreateSlide;
     'app-custom-images': AppCustomImages;
+    'app-dashboard': AppDashboard;
     'app-deck-or-slide': AppDeckOrSlide;
     'app-developer': AppDeveloper;
     'app-editor': AppEditor;
     'app-editor-actions': AppEditorActions;
     'app-editor-toolbar': AppEditorToolbar;
+    'app-faq': AppFaq;
     'app-feed': AppFeed;
     'app-feed-card': AppFeedCard;
     'app-feed-card-tags': AppFeedCardTags;