+++ webcomponents/inline-editor/CHANGELOG.md
@@ -1,3 +1,8 @@
+<a name="1.0.0-alpha.11"></a>
+# 1.0.0-alpha.11 (2019-07-30)
+
+* fix close image processing toolbar if document or any other elements is clicked
+
 <a name="1.0.0-alpha.10"></a>
 # 1.0.0-alpha.10 (2019-06-09)